API

MakieRepel.dist — Method

dist(a::VecTypes{2}, b::VecTypes{2})

Returns the Euclidean distance (2-distance) between the points in an efficient way.

MakieRepel.intersects — Method

intersects(a_origin::VecTypes{2}, a_widths::VecTypes{2}, b_origin::VecTypes{2}, b_widths::VecTypes{2})

Tests whether the two rectangles a and b intersect, returns boolean. Edges are considered to be within the rectangle.

MakieRepel.intersects — Method

intersects(origin::VecTypes{N}, widths::VecTypes{N}, point::VecTypes{N}) where N

Tests whether the point lies within the rectangle specified by origin and widths; returns a boolean. Edges are considered to be within the rectangle.

MakieRepel.repel_from_points — Function

repel_from_points(
    points::AbstractVector{<: Makie.VecTypes{2}}, boxes::AbstractVector{<: Rect2}, axisbbox::Rect2, niters = 10000; 
    padding = 4, x = true, y = true, halign = 0.5, valign = 0.5, data_radius = 5, selfpoint_radius = 3, 
    attraction = 1.9e-2, box_repulsion = 1.1e-2, point_repulsion = 0.99e-2
)::Vector{<: VecTypes{2, Float64}} # representing origins of bboxes

Arguments

points: The base points of data, in pixel space.

boxes: The bounding boxes of the texts, in pixel space.

axisbbox: The bounding box of the axis, usually Rect2f(Point2f(0), widths(scene.px_area[])). Used to ensure that text stays within the axis.

niters: The number of iterations for which to optimize. We don't test for convergence, so the loop will run for this many iterations every time.

Keyword arguments

padding: The padding around each box (usually in pixels).

x: Whether to repel in the x direction.

y: Whether to repel in the y direction.

halign: The horizontal alignment of the centroid of a box.

valign: The vertical alignment of the centroid of a box.

data_radius: The radius around each data point to be kept clear.

selfpoint_radius: The radius around the data point associated with the annotation to be kept clear. Usually less than data_radius.

attraction: The attractive force between a box and its associated point.

box_repulsion: The repulsive force between boxes.

point_repulsion: The repulsive force between boxes and data points.

MakieRepel.repulsivelabels! — Method

repulsivelabels(text::Vector{<: AbstractString}, positions::Vector{Point2f})

Implements repulsive labels a la ggrepel.

Attributes to tune repulsion

padding: The padding around each box (usually in pixels).

x: Whether to repel in the x direction.

y: Whether to repel in the y direction.

halign: The horizontal alignment of the centroid of a box.

valign: The vertical alignment of the centroid of a box.

data_radius: The radius around each data point to be kept clear.

selfpoint_radius: The radius around the data point associated with the annotation to be kept clear. Usually less than data_radius.

attraction: The attractive force between a box and its associated point.

box_repulsion: The repulsive force between boxes.

point_repulsion: The repulsive force between boxes and data points.

Extended help

Available attributes and their defaults

Available attributes and their defaults for MakieCore.Combined{MakieRepel.repulsivelabels!} are:

MakieRepel.repulsivelabels — Method

repulsivelabels(text::Vector{<: AbstractString}, positions::Vector{Point2f})

Implements repulsive labels a la ggrepel.

Attributes to tune repulsion

padding: The padding around each box (usually in pixels).

x: Whether to repel in the x direction.

y: Whether to repel in the y direction.

halign: The horizontal alignment of the centroid of a box.

valign: The vertical alignment of the centroid of a box.

data_radius: The radius around each data point to be kept clear.

selfpoint_radius: The radius around the data point associated with the annotation to be kept clear. Usually less than data_radius.

attraction: The attractive force between a box and its associated point.

box_repulsion: The repulsive force between boxes.

point_repulsion: The repulsive force between boxes and data points.

Extended help

Available attributes and their defaults

Available attributes and their defaults for MakieCore.Combined{MakieRepel.repulsivelabels} are:

  attraction        0.07f0
  box_repulsion     0.072f0
  cycle             [:color]
  data_radius       5
  font              :regular
  fonts             Attributes with 1 entry:
  regular => FTFont (family = TeX Gyre Heros Makie, style = Regular)
  fontsize          14
  halign            0.5
  linecolor         :black
  linestyle         :solid
  linevisible       true
  linewidth         1.0
  marker            :square
  markercolor       :black
  markersize        5
  niters            10000
  padding           10
  point_repulsion   0.051f0
  selfpoint_radius  7.5
  textalign         (:left, :bottom)
  textcolor         :black
  textrotation      0.0f0
  valign            0.5
  x                 true
  y                 true

MakieRepel.spring_repel — Method

spring_repel(
    a_origin::VecTypes{2}, a_widths::VecTypes{2}, 
    b_origin::VecTypes{2}, b_widths::VecTypes{2}; 
    k = 10000, x = true, y = true, halign = 0.5, valign = 0.5
)

Repels the centroids of the two given rectangle specifications a and b away from each other by Hooke's law. Returns the repulsion exerted by b on a.

If the rectangles intersect, their repulsion is multiplied by 5. If they do not intersect, it is muttiplied by 0.01.

halign and valign control the alignment of the centroid within the rectangle.

MakieRepel.spring_repel — Method

spring_repel(a_origin::VecTypes{2}, a_widths::VecTypes{2}, point::VecTypes{2})

Returns the spring-force repulsion between the centroid of the given rectangle and the point.

halign and valign control the alignment of the centroid within the rectangle.

MakieRepel.spring_repel — Method

spring_repel(a::VecTypes{2}, b::VecTypes{2}; k = 0.01, x = true, y = true)

Returns the repulsive force exerted on a by b, using Hooke's law: F = k⋅(a⃗-b⃗).

x and y represent whether to move in the x or y directions.