API

MakieRepel.dist — Method

dist(a::VecTypes{2}, b::VecTypes{2})

Returns the Euclidean distance (2-distance) between the points in an efficient way.

MakieRepel.intersects — Method

intersects(a_origin::VecTypes{2}, a_widths::VecTypes{2}, b_origin::VecTypes{2}, b_widths::VecTypes{2})

Tests whether the two rectangles a and b intersect, returns boolean. Edges are considered to be within the rectangle.

MakieRepel.intersects — Method

intersects(origin::VecTypes{N}, widths::VecTypes{N}, point::VecTypes{N}) where N

Tests whether the point lies within the rectangle specified by origin and widths; returns a boolean. Edges are considered to be within the rectangle.

MakieRepel.spring_repel — Method

spring_repel(
    a_origin::VecTypes{2}, a_widths::VecTypes{2}, 
    b_origin::VecTypes{2}, b_widths::VecTypes{2}; 
    k = 10000, x = true, y = true, halign = 0.5, valign = 0.5
)

Repels the centroids of the two given rectangle specifications a and b away from each other by Hooke's law. Returns the repulsion exerted by b on a.

If the rectangles intersect, their repulsion is multiplied by 5. If they do not intersect, it is muttiplied by 0.01.

halign and valign control the alignment of the centroid within the rectangle.

MakieRepel.spring_repel — Method

spring_repel(a_origin::VecTypes{2}, a_widths::VecTypes{2}, point::VecTypes{2})

Returns the spring-force repulsion between the centroid of the given rectangle and the point.

halign and valign control the alignment of the centroid within the rectangle.

MakieRepel.spring_repel — Method

spring_repel(a::VecTypes{2}, b::VecTypes{2}; k = 0.01, x = true, y = true)

Returns the repulsive force exerted on a by b, using Hooke's law: F = k⋅(a⃗-b⃗).

x and y represent whether to move in the x or y directions.